feature added - apply policies

+++ app/Http/Controllers/MyJobListingApplicationController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\JobListingApplication;
+use App\Policies\JobListingApplicationPolicy;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class MyJobListingApplicationController extends Controller {
     /**
@@ -26,6 +28,8 @@ class MyJobListingApplicationController extends Controller {
      * Remove the specified resource from storage.
      */
     public function destroy(JobListingApplication $myJobListingsApplication) {
+        Gate::authorize('delete', $myJobListingsApplication, JobListingApplication::class);
+
         $myJobListingsApplication->delete();
 
         return redirect()->back()->with('success', 'Job Application Cancelled.');
